Impl pwrite

+++ src/params.h
@@ -5,7 +5,7 @@
 namespace ulayfs {
 // hardware parameters
 constexpr static uint32_t BLOCK_SHIFT = 12;
-constexpr static uint32_t BLOCK_SIZE = 1 << BLOCK_SHIFT;
+constexpr static uint64_t BLOCK_SIZE = 1 << BLOCK_SHIFT;
 constexpr static uint32_t CACHELINE_SHIFT = 6;
 constexpr static uint32_t CACHELINE_SIZE = 1 << CACHELINE_SHIFT;
